refactor: add KnowledgeToFeedInput
