+++ frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/KnowledgeToFeed.tsx
@@ -1,13 +1,9 @@
-import { useTranslation } from "react-i18next";
 import { MdClose } from "react-icons/md";
 
 import Button from "@/lib/components/ui/Button";
-import { Divider } from "@/lib/components/ui/Divider";
 
-import { FeedItems } from "./components";
-import { Crawler } from "./components/Crawler";
-import { FileUploader } from "./components/FileUploader";
-import { FeedItemType, FeedItemUploadType } from "../../types";
+import { KnowledgeToFeedInput } from "./components/KnowledgeToFeedInput";
+import { FeedItemType } from "../../types";
 
 type FeedProps = {
   onClose: () => void;
@@ -21,12 +17,6 @@ export const KnowledgeToFeed = ({
   addContent,
   removeContent,
 }: FeedProps): JSX.Element => {
-  const { t } = useTranslation(["translation"]);
-
-  const files: File[] = (
-    contents.filter((c) => c.source === "upload") as FeedItemUploadType[]
-  ).map((c) => c.file);
-
   return (
     <div className="flex-col w-full relative">
       <div className="absolute right-2 top-1">
@@ -36,10 +26,12 @@ export const KnowledgeToFeed = ({
           </span>
         </Button>
       </div>
-      <FileUploader addContent={addContent} files={files} />
-      <Divider text={t("or")} className="m-5" />
-      <Crawler addContent={addContent} />
-      <FeedItems contents={contents} removeContent={removeContent} />
+
+      <KnowledgeToFeedInput
+        contents={contents}
+        addContent={addContent}
+        removeContent={removeContent}
+      />
     </div>
   );
 };
