+++ frontend/app/chat/[chatId]/components/ActionsBar/components/KnowledgeToFeed/components/KnowledgeToFeedInput/components/FeedItems/FeedItems.tsx
@@ -1,8 +1,8 @@
 import { Fragment } from "react";
 
+import { FeedItemType } from "../../../../../../types";
 import { CrawlFeedItem } from "./components/CrawlFeedItem";
 import { FileFeedItem } from "./components/FileFeedItem/FileFeedItem";
-import { FeedItemType } from "../../../../types";
 
 type FeedItemsProps = {
   contents: FeedItemType[];
